fix(test): move AmountFormattingTest with its subject, and make the locale scan run
Two CI failures on this PR, both real, and the second is the more interesting.

AmountFormattingTest reads three `internal` functions in NimazAmountInput.kt,
which moved to :core:ui. `internal` is module-scoped, so the test stopped
compiling. The 62 component tests migrated in the previous commit were all in
src/testDebug; this one is in src/test and was missed. It moves with its subject
and is repackaged to molecules, matching where NimazAmountInput actually lives.

DateTimeFormattingTest asserts SHIPPED_LOCALES matches the values-* directories
the app ships, reading them from app/src/main/res — which this PR emptied. The
test predicted exactly this in its own comment: the isNotEmpty assertion exists
so that a move becomes "a red build rather than a check that quietly stops
checking". It did its job. It now searches both roots so it survives the next
move too.

What the comment did not predict is that the failure would only appear on CI.
The task reads a directory in another module, which Gradle has no way to know is
an input, so moving the locales did not invalidate it: :core:common's tests
stayed UP-TO-DATE through two full local sweeps and only ran on CI, which starts
from an empty build directory.

That is the sharper form of the rule this epic keeps re-learning. Every guard
here has been verified by making it fail — but an assertion only fires if its
task runs, and none of the cross-module scans declare the external files they
read. core/common/build.gradle.kts now declares the resource directory as a test
input. Verified end to end rather than by inspection: `touch` proves nothing
because Gradle hashes content, so a real values-zz/ directory was added — the
task re-executed and the assertion failed — then removed, and it passed again.

Also re-ran :app and :core:ui unit-test compilation with --rerun-tasks, since
stale incremental state is precisely what hid both of these locally.

:core:ui 407 -> 419 tests, :app -12. A pure move; totals unchanged at 2402.
