+++ core/common/build.gradle.kts
@@ -1,3 +1,5 @@
+import org.gradle.api.tasks.PathSensitivity
+
 plugins {
     id("nimaz.android.library")
     // Merged into :app:jacocoTestReport via `coverageModules`. A module that leaves :app without
@@ -20,6 +22,18 @@ android {
 // reference `R` — a string that must be resolved outside a composable goes through
 // `StringProvider`, which is exactly why that interface exists. `coreCommonBoundary` enforces
 // both rules on every build.
+// `DateTimeFormattingTest` asserts that SHIPPED_LOCALES matches the `values-*` directories the app
+// actually ships — a fact that lives in another module's `res/`. Gradle cannot know that from the
+// source alone, so without this the task stays UP-TO-DATE when those directories change and the
+// assertion never runs. It did exactly that when PR 10 of #551 moved the five translations into
+// `:core:ui`: two full local sweeps reported the module green, and CI, which starts from an empty
+// build directory, failed. An assertion only fires if its task runs.
+tasks.withType<Test>().configureEach {
+    inputs.dir(rootProject.layout.projectDirectory.dir("core/ui/src/main/res"))
+        .withPropertyName("shippedLocaleResources")
+        .withPathSensitivity(PathSensitivity.RELATIVE)
+}
+
 dependencies {
     api(project(":core:domain"))
