+++ core/ui/src/test/kotlin/com/arshadshah/nimaz/presentation/components/molecules/AmountFormattingTest.kt
@@ -1,4 +1,4 @@
-package com.arshadshah.nimaz.presentation.components.atoms
+package com.arshadshah.nimaz.presentation.components.molecules
 
 import com.google.common.truth.Truth.assertThat
 import org.junit.Test
